+++ src/cqterrain/material/__init__.py
@@ -1,1 +1,4 @@
 from .stone import make_stones
+from .center_blocks import center_blocks
+from .uneven_blocks import uneven_blocks
+from .uneven_centered_blocks import uneven_centered_blocks